backend/requirements.txt
baseball_id
flask
flask-cors
flask-jwt-extended
flask-login
flask-sqlalchemy
joblib
matplotlib
pandas
plotly
python-dotenv
python-mlb-statsapi
scikit-learn==1.3.0
seaborn
